requirements.txt
amqp==5.1.1
asgiref==3.3.4
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
autopep8==1.5.4
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.21.27
botocore==1.24.27
celery==5.2.7
certifi==2020.12.5
cffi==1.14.3
channels==3.0.4
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==37.0.2
defusedxml==0.7.0rc1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2
django-braces==1.14.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.12.4
drf-social-oauth2==1.0.8
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==1.0.0
jwcrypto==0.8
kombu==5.2.4
mailjet-rest==1.3.4
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==21.3
Pillow==8.2.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.4
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.4.1
redis==4.3.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.2
service-identity==21.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.14.1
zope.interface==5.4.0
